+++ thrust_stand/PressureTransducer.cpp
@@ -7,7 +7,7 @@ PressureTransducer::PressureTransducer(uint8_t analogSignalPin)
     pinMode(analogSignalPin, INPUT);
 }
 
-double PressureTransducer::readBar() {
+float PressureTransducer::readBar() {
     int reading = analogRead(analogSignalPin);
     Serial.println(reading);
     //10bar / 5V * reading
